requirements.txt
discord.py>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.0

# dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0
